refactor: 수익률 계산 로직을 BettingAmount로 이동
- 베팅 금액에 대한 계산을 BetAmount에서 담당하도록 책임 이동

+++ src/main/java/domain/result/BettingResult.java
@@ -21,7 +21,7 @@ public class BettingResult {
         return participantName;
     }
 
-    public double getProfit() {
+    public long getProfit() {
         return bettingProfit.getProfit();
     }
 }
